refactor: move the request records to the request package
